Fixing linting error: File too long.
